Use synthetic projects for combinations
This commit completely revamps how the benchmark feature combinations work,
in order to use multiple projects, instead of having a single project
generate multiple artifacts.

In a nutshell, we define a "common" project which contains the shared
sources for all benchmarks. This project will be built with different
combinations of features, each combination resulting in a single project.

The combinations are declared in the `settings.gradle.kts` file, and
their behavior is defined via _convention plugins_. For example, if we
define a dimension called `epoll` with variants `on` and `off`, and
another dimension called `tcnative` with variants `on` and `off`, then
we will generate 4 synthetic projects:

   - :test-case:epoll-on:tcnative-on
   - :test-case:epoll-on:tcnative-off
   - :test-case:epoll-off:tcnative-on
   - :test-case:epoll-off:tcnative-off

When calling `shadowJar`, then all artifacts will
be produced, then copied into the root `build/libs` directory,
where we will find artifacts named against the combination,
for example `epoll-on-tcnative-off-all.jar`.

Similarly, `nativeCompile` will build all native images and
copy them into the root `build/native-images` directory,
with binaries named against the combination, for example
`epoll-on-tcnative-off`.

The behavior of each dimension variant must be defined in
a convention plugin, which will automatically be applied to
the project. For example, to configure tcnative "on", a
file named `io.micronaut.testcase.tcnative.on.gradle.kts`
must be found under `build-logic/src/main/kotlin`.

Not all variants have to provide such a plugin: only when
the build needs to be changed compared to the common build.

It is possible to generate a single artifact, say the native image,
for a single combination by using the full task path, for example:

```
./gradlew :test-case:tcnative-on:epoll-on:json-serde:nativeCompile
```

It is also possible to disable specific combinations, for example:

```gradle
configure<io.micronaut.bench.AppVariants> {
    combinations {
        dimension("tcnative") {
            variant("off")
            variant("on")
        }
        dimension("epoll") {
            variant("off")
            variant("on")
        }
        dimension("json") {
            variant("jackson")
            variant("serde")
        }
        exclude("tcnative-off/epoll-off/json-jackson")
    }
}
```

It is also possible to create independent combinations,
for example by declaring a 2d combinations block:

```gradle
combinations {
   dimension("tcnative") { ... }
   dimension("epoll") { ... }
}
combinations {
   dimension("json") { ... }
   dimension("loom") { ... }
}
```
```