Add support for Micronaut and Java versions
This commit updates the combination matrix to support configuring both
Micronaut 3 and 4 versions, as well as building with different Java
versions.

In a nutshell, it is now possible to have specific source directories
per "variant name". For example, common Java sources are found in
`src/main/java`, but for the `java-11` variant, then the project
will _also_ look into `src/main/variants/java-11/java`. Similarly,
resources can be included into `src/main/variants/java-11/resources`.

Support for Micronaut 4 requires depending on the snapshot version
of the Micronaut Gradle plugin, and tweaks to build Micronaut 3
applications, which are implemented in the corresponding feature
plugin.